added the employee record

+++ src/main/java/com/example/mongoEmployee/models/EmployeeModel.java
@@ -2,5 +2,5 @@ package com.example.mongoEmployee.models;
 
 import java.math.BigDecimal;
 
-public record EmployeeModel(String name, int age, BigDecimal salary) {
+public record EmployeeModel(String name, int age, Double salary) {
 }
